+++ apps/extension/src/routes/popup/home/block-sync.tsx
@@ -1,11 +1,11 @@
-import { BlockSyncStatus } from '@penumbra-zone/ui';
+import { LargeBlockSyncStatus } from '@penumbra-zone/ui';
 import { useSyncProgress } from '../../../hooks/full-sync-height';
 
 export const BlockSync = () => {
   const { latestBlockHeight, fullSyncHeight } = useSyncProgress();
 
   return (
-    <BlockSyncStatus
+    <LargeBlockSyncStatus
       fullSyncHeight={fullSyncHeight}
       latestKnownBlockHeight={latestBlockHeight}
       showLoadingState={false}
