+++ packages/ui/components/ui/block-sync-status/shared.ts
@@ -1,0 +1,10 @@
+export interface BlockSyncProps {
+  latestKnownBlockHeight?: number;
+  fullSyncHeight?: number;
+  error?: unknown;
+}
+
+export interface SyncingStateProps {
+  latestKnownBlockHeight: number;
+  fullSyncHeight: number;
+}
